Move interpolation operations to another module
